Add environment variables to a file for cron to access properly
The issue was that the script was not working when run as a cron job,
but did work when run in bash.

The reason for this is that cron runs in a separate environment  where $USERNAME
and $PASSWORD are not set.
The fix to this was to explicitly give cron the correct environment,
see this
[SO Q/A](https://stackoverflow.com/questions/27771781/how-can-i-access-docker-set-environment-variables-from-a-cron-job)